several changes
Abandoned the approach of creating a list of unique characters and then
iterating on that list. Removed all code related to that.

Instead I begin with an empty dictionary, then create a list containing
all the words or strings in my file, I then look at each character in
each word and add them to the dictionary while implementing a counter
to count each time an item or character appears.